clean up and make hydrate the source of truth
